requirements.txt
# Base dependencies - Python 3.12 compatible
numpy>=1.26.0

# Deep learning - Python 3.12 compatible
tensorflow>=2.16.0

# Image processing
Pillow>=11.1.0
opencv-python>=4.8.0

# Web framework and utilities
Flask>=2.0.2
flask-cors==4.0.1
Flask-SQLAlchemy>=3.0.0
python-dotenv>=1.1.0
gunicorn>=21.0.0

# Face detection
--extra-index-url https://pypi.org/simple/
fire==0.7.0
deepface>=0.0.95

# Additional dependencies for deepface
mtcnn>=0.1.0
retina-face>=0.0.1
gdown>=3.10.1
tqdm>=4.30.0
pandas>=2.0.0
openpyxl>=3.1.2
boto3>=1.26.0
openai
pusher

modal>=1.0.0
anthropic

# Vector database for face embeddings
psycopg2-binary>=2.9.9
pgvector>=0.2.4
